% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CausalMetricSpaces-package.R
\docType{package}
\name{CausalMetricSpaces-package}
\alias{CausalMetricSpaces}
\alias{CausalMetricSpaces-package}
\title{CausalMetricSpaces: What the Package Does (Title Case)}
\description{
More about what it does (maybe more than one line) Use four spaces when indenting paragraphs within the Description.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/herbps10/CausalMetricSpaces}
  \item Report bugs at \url{https://github.com/herbps10/CausalMetricSpaces/issues}
}

}
\keyword{internal}
